# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chamber Commissions API — سامانه کمیسیون‌های تخصصی
  version: 1.0.0
  description: 'REST API of the commissions platform. All business endpoints require `Authorization: Bearer <accessToken>`.


    Real-time: Socket.IO at `/socket.io` (auth: `{ token }`). Emit `meeting:join` with a meeting id, then receive `meeting.updated`, `attendance.updated`
    (officers), `quorum.updated`, `agenda.updated`, `comment.created`, `vote.opened`, `vote.progress`, `vote.closed`, and personal `notification`
    events.


    Errors: `{ error: { code, message, details? } }` with Persian messages.'
servers:
- url: http://localhost:4000
security:
- bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: &id001
              type: string
            message: *id001
            details:
              type: array
              items:
                type: object
                properties:
                  path: *id001
                  message: *id001
...
paths:
...
  /api/documents:
    post:
      tags:
      - Documents
      summary: 'بارگذاری سند (multipart: file + meetingId|agendaItemId|resolutionId|issueId|commissionId)'
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: Type/extension/magic-byte validated, size limited, malware scan hook.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                meetingId: *id004
                agendaItemId: *id004
                resolutionId: *id004
                issueId: *id004
                commissionId: *id004
                kind: *id001
                title: *id001
    get:
      tags:
      - Documents
      summary: فهرست اسناد
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/documents/{id}/download:
    get:
      tags:
      - Documents
      summary: دریافت سند
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/documents/{id}/link:
    post:
      tags:
      - Documents
      summary: لینک امضاشده کوتاه‌مدت برای موبایل
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
  /api/files/{id}:
    get:
      tags:
      - Documents
      summary: دریافت سند با لینک امضاشده (?token) — بدون Bearer
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '404':
          description: Not found or not visible
        '409':
          description: Conflict (state machine / business rule)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      description: Token valid for 2 minutes; access is re-checked.
      security: []
      parameters:
      - name: id
        in: path
        required: true
        schema: *id004
...